requirements.txt
bcrypt==5.0.0
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==2.0.0
cssselect2==0.8.0
defusedxml==0.7.1
Faker==37.11.0
pillow==11.3.0
pycparser==2.23
pyglet==2.1.9
tinycss2==1.4.0
tzdata==2025.2
watchdog==6.0.0
webencodings==0.5.1
